The issues with numerical derivation for the local energy solved. The error lay in the implementation of the evaluation of the wavefunction.
Changed the evaluation function to take into consideration every
particle in the system and implemented the change over the program.

Additionally, rectified the expression for the derivation to be scaled
properly with the steplength.

the results produced now follow expectations of energies both for
initial energy of 1 particle and 1 dimension as well as scaling for
either.